requirements.txt
# Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0

# Authentication
authlib==1.4.0
itsdangerous==2.2.0
python-jose[cryptography]==3.3.0
httpx==0.28.1

# Transbank
transbank-sdk==5.0.0

# Validation and config
pydantic==2.10.4
pydantic-settings==2.7.0
email-validator==2.2.0

# Templates
jinja2==3.1.5

# Email
aiosmtplib==3.0.2

# Utils
python-dotenv==1.0.1
